pyproject.toml
[project]
name = "nhm-django-infra"
version = "1.0.0"
description = "A Django-based infrastructure project providing backend services for the Porsche application, featuring REST API support, PostgreSQL database integration, and development workflow automation tools."
authors = [
    { name = "xuheng.zong", email = "79689669+nohairman@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.0",
    "dynaconf>=3.2.7",
    "djangorestframework>=3.15.2",
    "django-filter>=24.3",
    "gunicorn>=23.0.0",
    "psycopg[binary,pool]>=3.2.9",
    "ujson>=5.11.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.uv.sources]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.black]
line_length = 120

[tool.isort]
profile = "black"
src_paths = ["src"]
py_version = 313
skip_gitignore = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["porsche"]
multi_line_output = 3
line_length = 120
